[#21] Rename: apis 하위 파일명 변경
apis 폴더 하위의 파일들은 디렉토리 명이 apis이므로 ~Api와 같이 표시하지 않아도 이해할 수 있다고 생각하여 이름을 변경하였습니다.
또 signinApi 안에 postSignUp을 포함하고 있어 해당 api 파일의 명칭을 auth로 변경하였습니다.
파일명의 변경으로 인해 import 경로 부분이 수정되었습니다.
